Add tricky list parsing instances
RTM has a really backwards, bad way of exposing lists. For contacts, the
contacts list will be found under the object keys "objects" then under
"object". The type if there is one is an object, and if there are more,
it is an Array. Weak. Therefor, we have to handle the object case by
wrapping it in an Array (Vector).